Added admin settings & unistall
Admin settings and uninstall added to plugin. When unistalling all
entries in the options table are removed

+++ uninstall.php
@@ -3,13 +3,12 @@
 	// If uninstall not called from WordPress exit
 
 	if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {
-	
 		exit();
-	
 	}
 	
 	// Delete settings page options from options table
 	
-	delete_option( 'gyn_options' );
-	
+	if ( get_option( 'gyn_options' ) != false ) {
+		delete_option( 'gyn_options' );
+	}
 ?>
